+++ clang/lib/Parse/ParseExprCXX.cpp
@@ -2040,7 +2040,8 @@ Parser::ParseAliasDeclarationInInitStatement(DeclaratorContext Context,
 Sema::ConditionResult
 Parser::ParseCXXCondition(StmtResult *InitStmt, SourceLocation Loc,
                           Sema::ConditionKind CK, bool MissingOK,
-                          ForRangeInfo *FRI, bool EnterForConditionScope) {
+                          ForRangeInfo *FRI, bool EnterForConditionScope,
+                          SourceLocation ConstexprLoc) {
   // Helper to ensure we always enter a continue/break scope if requested.
   struct ForConditionScopeRAII {
     Scope *S;
@@ -2097,9 +2098,28 @@ Parser::ParseCXXCondition(StmtResult *InitStmt, SourceLocation Loc,
       *InitStmt = Actions.ActOnNullStmt(SemiLoc);
       return ParseCXXCondition(nullptr, Loc, CK, MissingOK);
     }
+    bool InitStmtIsExprStmt = false;
+    if (InitStmt) {
+      RevertingTentativeParsingAction PA(*this);
+      SkipUntil(tok::r_paren, tok::semi, StopBeforeMatch);
+      InitStmtIsExprStmt = Tok.is(tok::semi);
+    }
 
-    // Parse the expression.
-    ExprResult Expr = ParseExpression(); // expression
+    ExprResult Expr; // expression
+    {
+      EnterExpressionEvaluationContext Consteval(
+          Actions, Sema::ExpressionEvaluationContext::ConstantEvaluated,
+          /*LambdaContextDecl=*/nullptr,
+          Sema::ExpressionEvaluationContextRecord::EK_Other,
+          /*ShouldEnter=*/CK == Sema::ConditionKind::ConstexprIf &&
+              !InitStmtIsExprStmt);
+      SourceLocation OuterConstexprIfLoc = Actions.ConstexprIfLoc;
+      Actions.ConstexprIfLoc = ConstexprLoc;
+
+      // Parse the expression.
+      Expr = ParseExpression(); // expression
+      Actions.ConstexprIfLoc = OuterConstexprIfLoc;
+    }
     if (Expr.isInvalid())
       return Sema::ConditionError();
 
@@ -2187,6 +2207,21 @@ Parser::ParseCXXCondition(StmtResult *InitStmt, SourceLocation Loc,
   if (CopyInitialization)
     ConsumeToken();
 
+  Sema::ExpressionEvaluationContext NewEEC =
+      Actions.ExprEvalContexts.back().Context;
+  bool RuntimeEvaluated = Actions.ExprEvalContexts.back().IsRuntimeEvaluated;
+  if (DS.getTypeQualifiers() == DeclSpec::TQ_const)
+    RuntimeEvaluated = false;
+
+  if (CK == Sema::ConditionKind::ConstexprIf || DS.hasConstexprSpecifier()) {
+    RuntimeEvaluated = false;
+    NewEEC = Sema::ExpressionEvaluationContext::ConstantEvaluated;
+  }
+  EnterExpressionEvaluationContext Initializer(Actions, NewEEC, DeclOut);
+  Actions.ExprEvalContexts.back().IsRuntimeEvaluated = RuntimeEvaluated;
+  SourceLocation OuterConstexprIfLoc = Actions.ConstexprIfLoc;
+  Actions.ConstexprIfLoc = ConstexprLoc;
+
   ExprResult InitExpr = ExprError();
   if (getLangOpts().CPlusPlus11 && Tok.is(tok::l_brace)) {
     Diag(Tok.getLocation(),
@@ -2213,6 +2248,7 @@ Parser::ParseCXXCondition(StmtResult *InitStmt, SourceLocation Loc,
     Actions.ActOnInitializerError(DeclOut);
 
   Actions.FinalizeDeclaration(DeclOut);
+  Actions.ConstexprIfLoc = OuterConstexprIfLoc;
   return Actions.ActOnConditionVariable(DeclOut, Loc, CK);
 }